requirements.txt
dj-database-url
Django
gunicorn
pytz
PyYAML
whitenoise
mysql
pillow
django-tinymce
boto
django-storages